Psalm level 3

See merge request eternal-twin/mush!104

+++ Api/src/Action/Actions/GetUp.php
@@ -44,11 +44,10 @@ class GetUp extends Action
 
     protected function applyEffects(): ActionResult
     {
-        $lyingDownStatus = $this->player->getStatusByName(PlayerStatusEnum::LYING_DOWN);
-
-        $lyingDownStatus->setPlayer(null)->setGameEquipment(null);
-
-        $this->statusService->delete($lyingDownStatus);
+        if ($lyingDownStatus = $this->player->getStatusByName(PlayerStatusEnum::LYING_DOWN)) {
+            $lyingDownStatus->setPlayer(null)->setGameEquipment(null);
+            $this->statusService->delete($lyingDownStatus);
+        }
 
         return new Success();
     }
